+++ packages/inferencer/src/inferencers/headless/create.tsx
@@ -1,5 +1,4 @@
-import * as RefineCore from "@refinedev/core";
-import * as RefineReactHookForm from "@refinedev/react-hook-form";
+import { useForm } from "@refinedev/react-hook-form";
 
 import { createInferencer } from "@/create-inferencer";
 import {
@@ -345,12 +344,7 @@ export const renderer = ({
 export const CreateInferencer: InferencerResultComponent = createInferencer({
     type: "create",
     additionalScope: [
-        ["@refinedev/core", "RefineCore", RefineCore],
-        [
-            "@refinedev/react-hook-form",
-            "RefineReactHookForm",
-            RefineReactHookForm,
-        ],
+        ["@refinedev/react-hook-form", "RefineReactHookForm", { useForm }],
     ],
     codeViewerComponent: SharedCodeViewer,
     loadingComponent: LoadingComponent,