+++ packages/inferencer/src/inferencers/mui/list.tsx
@@ -1,9 +1,18 @@
-import * as RefineMui from "@refinedev/mui";
-import * as EmotionReact from "@emotion/react";
-import * as EmotionStyled from "@emotion/styled";
-import * as MuiLab from "@mui/lab";
-import * as MuiMaterial from "@mui/material";
-import * as MuiXDataGrid from "@mui/x-data-grid";
+import {
+    useDataGrid,
+    EditButton,
+    SaveButton,
+    DeleteButton,
+    List,
+    TagField,
+    EmailField,
+    UrlField,
+    DateField,
+    MarkdownField,
+    ShowButton,
+} from "@refinedev/mui";
+import { DataGrid } from "@mui/x-data-grid";
+import { Checkbox } from "@mui/material";
 
 import { createInferencer } from "@/create-inferencer";
 import {
@@ -771,12 +780,25 @@ export const renderer = ({
 export const ListInferencer: InferencerResultComponent = createInferencer({
     type: "list",
     additionalScope: [
-        ["@refinedev/mui", "RefineMui", RefineMui],
-        ["@emotion/react", "EmotionReact", EmotionReact],
-        ["@emotion/styled", "EmotionStyled", EmotionStyled],
-        ["@mui/lab", "MuiLab", MuiLab],
-        ["@mui/material", "MuiMaterial", MuiMaterial],
-        ["@mui/x-data-grid", "MuiXDataGrid", MuiXDataGrid],
+        [
+            "@refinedev/mui",
+            "RefineMui",
+            {
+                useDataGrid,
+                EditButton,
+                SaveButton,
+                DeleteButton,
+                List,
+                TagField,
+                EmailField,
+                UrlField,
+                DateField,
+                MarkdownField,
+                ShowButton,
+            },
+        ],
+        ["@mui/x-data-grid", "MuiXDataGrid", { DataGrid }],
+        ["@mui/material", "MuiMaterial", { Checkbox }],
     ],
     codeViewerComponent: SharedCodeViewer,
     loadingComponent: LoadingComponent,