+++ packages/mantine/src/components/buttons/export/index.tsx
@@ -1,6 +1,9 @@
 import React from "react";
 import { useTranslate } from "@refinedev/core";
-import { RefineButtonTestIds } from "@refinedev/ui-types";
+import {
+    RefineButtonClassNames,
+    RefineButtonTestIds,
+} from "@refinedev/ui-types";
 import { ActionIcon, Button } from "@mantine/core";
 import { IconFileExport } from "@tabler/icons";
 
@@ -33,6 +36,7 @@ export const ExportButton: React.FC<ExportButtonProps> = ({
                 : { variant: "default" })}
             loading={loading}
             data-testid={RefineButtonTestIds.ExportButton}
+            className={RefineButtonClassNames.ExportButton}
             {...commonProps}
         >
             <IconFileExport size={18} {...svgIconProps} />
@@ -43,6 +47,7 @@ export const ExportButton: React.FC<ExportButtonProps> = ({
             loading={loading}
             leftIcon={<IconFileExport size={18} {...svgIconProps} />}
             data-testid={RefineButtonTestIds.ExportButton}
+            className={RefineButtonClassNames.ExportButton}
             {...rest}
         >
             {children ?? translate("buttons.export", "Export")}